[refactor] #50 response 파일명 오타 변경
